Added Local splitscreen/Multiplayer
Added blueprints for Local multiplayer/splitscreen.
Tested on my setup with one keyboard and one gamepad , it worked as intended.. Keypress from gamepad splits the screen and creates a new player that has gamepad as control.

I tried to plug in 4 different controllers and the game split into 4 , so it works with 4 players at the moment